+++ code/src/sixsq/nuvla/server/resources/session_password.clj
@@ -80,10 +80,8 @@ password.
         cookie      (cookies/create-cookie cookie-info)
         expires     (ts/rfc822->iso8601 (:expires cookie))
         claims      (:claims cookie-info)
-        groups      (:groups cookie-info)
         session     (cond-> (assoc session :expiry expires)
-                            claims (assoc :roles claims)
-                            groups (assoc :groups groups))
+                            claims (assoc :roles claims))
         cookies     {authn-info/authn-cookie cookie}]
     (log/debug "password cookie token claims for"
                (u/id->uuid href) ":" cookie-info)