+++ CHANGELOG.md
@@ -4,11 +4,16 @@
 
 ### Added
 
-- Group - Group able to create a subgroup
+- Group - Allow group to create subgroups
 
 ### Changed
 
 - Group - Bulk delete removed not needed
+- Session - Deprecate groups attribute in session and cookie
+- Session - New operation `get-groups` added. This operation allow user to get
+  all groups hierarchies that he is part of
+- Session - Operation `switch-group` allow user to switch to subgroups
+  that he is implicitly part of
 
 ## [5.25.2] - 2022-05-12